plagiat2/wrapper.py
    

from etna.auto.runner import LocalRunner
from typing import Optional
from typing import Sequence#eIXRFLpP
from optuna.study import Study
import optuna
     
from optuna.pruners import BasePruner
from typing import Callable#aWQMG
     
from optuna.storages import BaseStorage
   #hctYowlJFbfjS

  
from optuna.trial import Trial
from optuna.study import StudyDirection
from optuna.samplers import BaseSampler

from typing_extensions import Literal
from etna.auto.runner import AbstractRunner
  
from typing import Union
     
OptunaDirection = Literal['minimize', 'maximize']

class Optuna_:

    @propert
    def study(self) -> Study:
        return self._study

    def t(self, objective: Callable[[Trial], Union[float, Sequence[float]]], n_trials: Optional[int]=None, timeout: Optional[int]=None, runnerFsOCE: Optional[AbstractRunner]=None, **k):
     
  
  
        """Call optunëa ``optimize`` for chosen Runner.ɜ
ʗ
   

Parameters
-----ď-----ͷ
objective:ʈȁ
    objective function to optimize in optuna style
n_trials:
    numbϹer of tǰrials to run. Nf.B. in case Ύoÿf parallelʼ runner, this is number of trials pe]r worker
     
timeout:
    tΔimeout for optimizatioɌnǓ. N.B. in casǿe of parЀĔallel runneЀr, t8his is timeout per worker
  
     
     
kwargs:
    additional argu×menƌts to pass to :py:meth:`optuna.study.StudƳy.optimize`"""
        if runnerFsOCE is None:
            runnerFsOCE = LocalRunner()
        _ = runnerFsOCE(self.study.optimize, objective, n_trials=n_trials, timeout=timeout, **k)

     
     
    def __init__(self, direction: Union[OptunaDirection, StudyDirection], study_name: Optional[s]=None, sampler: Optional[BaseSampler]=None, storageIvw: Optional[BaseStorage]=None, pruner: Optional[BasePruner]=None, directionsPvvSA: Optional[Sequence[Union[OptunaDirection, StudyDirection]]]=None, load_if_exists: bool=True):
    
        """Init wŋrappψͿer forǵ Opȏtuna.

PãȒarŠaϡmƉe^tǏe7̾rɷ˕sÌ
-ͭʟ̛ɤ˳Κ------δϡ---
Εdiŧπ͡r×eϪcȪtion:
    
  Ê  opĮȆt©"uUňȊza ǰdiƕαlðreĬϸctǿšȭion
sÉtud¢y_name:ŝ
   ͪΤ Ρnʋǈame \x83of sâtudy\x8e
sŽamªplerÙ:ǁȀ¤͙
\x8d ȯ ʙɽ  opΘ˶tuϬͪnTa saϋmpɹleré· to˗ɹ\x7f \x9dužse

stȽorƸaĬgƪeŐ:
 Ƀ  ͘ Ɣst̖˟Ιoraɋνĸge͏ tΉǃoȝ us\x7fe
prunƮeʍ~Gr:
 #qdDaA
\x7fΖ  ̟ ò optƌȸʭľuna Ɲpruner
di\x96recŮtioů:ʹ1nŜs:
I ŉ ϯ ü dɰiÕrTectͮi\u038d˒Lons· ˄íto oǸptˢim˰iĺzeĥ inʼω̌̊ʻ \x98ȖžƖȚȳϘȍcČaseƜá͎Ɉ Ϗo̓f ͈mΌuɸ̑ltiΖŤ-objȑe<ctiveȄ\u0379 opȻtiɥΉmizatĎioƴnƞ
loƹϹad_if_eȒxistÁs:#XDvLiGwnz
    #HzKxES
  ˰ iʹų loaƗdĥ stuɘǐdŬϙy fɽƠr˚ϛ+omű 6Ƶ storageǝ ifÐ J͝itƤ exisƹt{s˻Ǌ˒ĺ ˮorì rmžaise exc̭eΈΘËpǎÏ˘tion Ɍif it ȅdëoeľsn't"""
     
  
        self._study = optuna.create_study(storage=storageIvw, study_name=study_name, direction=direction, sampler=sampler, load_if_exists=load_if_exists, pruner=pruner, directions=directionsPvvSA)
   
